Switching to video.js-dist
The VideoJS bower package doesn't currently come with a dist folder that
I can commit without hacking on the bower repo. I'd rather not do that,
so I'll use this 3rd party video.js dist that has the files we need.
This is importnant because bower isn't easy to get running on our
Windows setup.